backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
pydantic-settings==2.7.1
python-jose==3.4.0
passlib==1.7.4
httpx==0.27.2
python-multipart==0.0.22
alembic==1.14.0
email-validator==2.2.0
pgvector==0.3.6
starlette==0.49.1
redis[hiredis]>=5.0
